examples/crytocurrency/requirements.txt
websocket-client
seaborn
pandas
requests
ipykernel
bokeh
scikit-learn